docs/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pgvector
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
sse-starlette
boto3
alembic
